fixed conflicts

+++ src/TheThingsNetwork.h
@@ -51,7 +51,7 @@ class TheThingsNetwork
     String readValue(String key);
     bool sendCommand(String cmd);
     bool sendCommand(String cmd, String value);
-    bool sendCommand(String cmd, const byte *buf, int length);
+    bool sendCommand(String cmd, const byte *buf, size_t length);
     void reset(bool adr = true);
     void configureEU868(uint8_t sf);
     void configureUS915(uint8_t sf, uint8_t fsb);
